test(sandbox): run provider-agnostic sandbox tests in the core suite
Split tests/unit_tests/test_sandbox.py so the provider-agnostic core
sandbox API tests (facade, registry, spec validation, sync loop runner,
mini-SWE sandbox environment) no longer carry the 'sandbox' pytest
marker and run in the default core unit test suite, which needs no
optional dependencies.

The OpenSandbox-specific tests move to test_opensandbox_provider.py,
which already guards on the optional 'sandbox' extra via importorskip
and keeps the 'sandbox' marker, so provider tests keep running in the
dedicated CI step (pytest -m sandbox) exactly as before.

No CI workflow changes needed: the unmarked core tests are picked up
automatically by ng_dev_test / -m 'not sandbox', and -m sandbox still
selects the opensandbox/docker/apptainer provider tests.
