+++ tests/unit_tests/test_sandbox.py
@@ -13,9 +13,7 @@
 # limitations under the License.
 
 import asyncio
-import importlib.util
 import threading
-from datetime import timedelta
 from pathlib import Path
 from types import SimpleNamespace
 from typing import Any
@@ -27,7 +25,6 @@ import nemo_gym.sandbox.providers.registry as provider_registry
 from nemo_gym.sandbox import (
     AsyncSandbox,
     Sandbox,
-    SandboxCreateError,
     SandboxExecResult,
     SandboxHandle,
     SandboxResources,
@@ -43,41 +40,6 @@ from nemo_gym.sandbox import (
 from nemo_gym.sandbox.api import _AsyncLoopRunner
 from nemo_gym.sandbox.utils import rewrite_image
 from responses_api_agents.mini_swe_agent_2.sandbox_environment import MiniSWESandboxEnvironment
-
-
-pytestmark = pytest.mark.sandbox
-
-
-def _has_module(module_name: str) -> bool:
-    try:
-        return importlib.util.find_spec(module_name) is not None
-    except ModuleNotFoundError:
-        return False
-
-
-requires_tenacity = pytest.mark.skipif(
-    not _has_module("tenacity"),
-    reason="tenacity optional sandbox dependency is not installed",
-)
-
-
-def _require_opensandbox_provider() -> tuple[Any, Any, Any, str, str]:
-    pytest.importorskip("tenacity", reason="tenacity optional sandbox dependency is not installed")
-    from nemo_gym.sandbox.providers.opensandbox import provider as opensandbox_provider_module
-    from nemo_gym.sandbox.providers.opensandbox.provider import (
-        IMAGE_PULL_POLICY_ANNOTATION_EXTENSION_KEY,
-        IMAGE_PULL_POLICY_EXTENSION_KEY,
-        OpenSandboxCreateVerificationError,
-        OpenSandboxProvider,
-    )
-
-    return (
-        opensandbox_provider_module,
-        OpenSandboxProvider,
-        OpenSandboxCreateVerificationError,
-        IMAGE_PULL_POLICY_EXTENSION_KEY,
-        IMAGE_PULL_POLICY_ANNOTATION_EXTENSION_KEY,
-    )
 
 
 class FakeSandboxProvider:
@@ -682,404 +644,6 @@ def test_sync_sandbox_facade_rejects_async_context() -> None:
         assert "use AsyncSandbox in async code" in str(e)
     else:
         raise AssertionError("expected sync Sandbox to reject async context")
-
-
-@requires_tenacity
-def test_opensandbox_sdk_create_receives_default_image_pull_policy(monkeypatch) -> None:
-    asyncio.run(_assert_opensandbox_sdk_create_receives_default_image_pull_policy(monkeypatch))
-
-
-async def _assert_opensandbox_sdk_create_receives_default_image_pull_policy(monkeypatch) -> None:
-    (
-        opensandbox_provider_module,
-        OpenSandboxProvider,
-        _OpenSandboxCreateVerificationError,
-        IMAGE_PULL_POLICY_EXTENSION_KEY,
-        IMAGE_PULL_POLICY_ANNOTATION_EXTENSION_KEY,
-    ) = _require_opensandbox_provider()
-    del _OpenSandboxCreateVerificationError
-
-    class FakeSDKSandbox:
-        create_calls: list[dict[str, Any]] = []
-
-        def __init__(self, sandbox_id: str) -> None:
-            self.id = sandbox_id
-
-        @classmethod
-        async def create(cls, **kwargs: Any) -> "FakeSDKSandbox":
-            cls.create_calls.append(kwargs)
-            return cls("sdk-sandbox-1")
-
-    monkeypatch.setattr(
-        opensandbox_provider_module,
-        "_require_opensandbox_sdk",
-        lambda: (FakeSDKSandbox, object, object, object, object),
-    )
-
-    provider = OpenSandboxProvider(probe={"command": None})
-    monkeypatch.setattr(provider, "_connection_config", lambda request_timeout_s=None: object())
-
-    handle = await provider.create(
-        SandboxSpec(
-            image="image:tag",
-            metadata={
-                "harbor_instance_id": "swebench::django__django-10880",
-                "long": f"bad:{'x' * 80}:",
-            },
-        )
-    )
-
-    assert handle.sandbox_id == "sdk-sandbox-1"
-    metadata = FakeSDKSandbox.create_calls[0]["metadata"]
-    assert metadata["harbor_instance_id"] == "swebench_django__django-10880"
-    assert metadata["long"] == ("bad_" + "x" * 59)
-    extensions = FakeSDKSandbox.create_calls[0]["extensions"]
-    assert extensions[IMAGE_PULL_POLICY_EXTENSION_KEY] == "IfNotPresent"
-    assert extensions[IMAGE_PULL_POLICY_ANNOTATION_EXTENSION_KEY] == "IfNotPresent"
-
-
-@requires_tenacity
-def test_opensandbox_connect_after_create_preserves_request_timeout(monkeypatch) -> None:
-    asyncio.run(_assert_opensandbox_connect_after_create_preserves_request_timeout(monkeypatch))
-
-
-async def _assert_opensandbox_connect_after_create_preserves_request_timeout(monkeypatch) -> None:
-    opensandbox_provider_module, OpenSandboxProvider, *_unused = _require_opensandbox_provider()
-
-    class FakeConnectionConfig:
-        def __init__(self, **kwargs: Any) -> None:
-            self.kwargs = kwargs
-
-    class FakeSDKSandbox:
-        connect_calls: list[dict[str, Any]] = []
-
-        def __init__(self, sandbox_id: str) -> None:
-            self.id = sandbox_id
-
-        @classmethod
-        async def connect(cls, sandbox_id: str, **kwargs: Any) -> "FakeSDKSandbox":
-            cls.connect_calls.append({"sandbox_id": sandbox_id, **kwargs})
-            return cls(sandbox_id)
-
-    monkeypatch.setattr(
-        opensandbox_provider_module,
-        "_require_opensandbox_sdk",
-        lambda: (FakeSDKSandbox, FakeConnectionConfig, object, object, object),
-    )
-
-    provider = OpenSandboxProvider(
-        connection={"domain": "sandbox.example", "protocol": "https", "request_timeout_s": 300},
-        create={"connect_attempt_timeout_s": 1},
-        probe={"command": None},
-    )
-    handle = await provider._connect_after_create(
-        SandboxHandle(sandbox_id="sdk-sandbox-1", provider_name="opensandbox", raw=None),
-        SandboxSpec(image="image:tag", ready_timeout_s=10),
-    )
-
-    assert handle.sandbox_id == "sdk-sandbox-1"
-    assert isinstance(handle.raw, FakeSDKSandbox)
-    connect_call = FakeSDKSandbox.connect_calls[0]
-    assert connect_call["skip_health_check"] is True
-    assert connect_call["connection_config"].kwargs == {
-        "domain": "sandbox.example",
-        "protocol": "https",
-        "request_timeout": timedelta(seconds=300),
-    }
-
-
-@requires_tenacity
-def test_opensandbox_create_probe_can_require_stable_successes(monkeypatch) -> None:
-    asyncio.run(_assert_opensandbox_create_probe_can_require_stable_successes(monkeypatch))
-
-
-async def _assert_opensandbox_create_probe_can_require_stable_successes(monkeypatch) -> None:
-    _opensandbox_provider_module, OpenSandboxProvider, *_unused = _require_opensandbox_provider()
-
-    provider = OpenSandboxProvider(
-        probe={
-            "command": "true",
-            "expected_stdout": None,
-            "stable_count": 3,
-            "stable_delay_s": 0,
-        },
-    )
-    calls: list[dict[str, Any]] = []
-
-    async def fake_exec(
-        handle: SandboxHandle,
-        command: str,
-        *,
-        cwd: str | None = None,
-        env: dict[str, str] | None = None,
-        timeout_s: int | float | None = None,
-        user: str | int | None = None,
-    ) -> SandboxExecResult:
-        calls.append(
-            {
-                "handle": handle,
-                "command": command,
-                "cwd": cwd,
-                "env": env,
-                "timeout_s": timeout_s,
-                "user": user,
-            }
-        )
-        return SandboxExecResult(stdout="", stderr="", return_code=0)
-
-    monkeypatch.setattr(provider, "_exec", fake_exec)
-    handle = SandboxHandle(sandbox_id="sdk-sandbox-0", provider_name="opensandbox", raw=object())
-
-    await provider._verify_created_handle(handle)
-
-    assert [call["command"] for call in calls] == ["true", "true", "true"]
-    assert all(call["timeout_s"] == 30 for call in calls)
-    assert all(call["user"] == "root" for call in calls)
-
-
-@requires_tenacity
-def test_opensandbox_create_probe_polls_same_sandbox_after_transient_errors(monkeypatch) -> None:
-    asyncio.run(_assert_opensandbox_create_probe_polls_same_sandbox_after_transient_errors(monkeypatch))
-
-
-async def _assert_opensandbox_create_probe_polls_same_sandbox_after_transient_errors(monkeypatch) -> None:
-    _opensandbox_provider_module, OpenSandboxProvider, *_unused = _require_opensandbox_provider()
-
-    provider = OpenSandboxProvider(
-        create={"connect_poll_s": 0.01},
-        probe={
-            "command": "true",
-            "expected_stdout": None,
-            "timeout_s": 1,
-            "deadline_s": 2,
-            "stable_count": 2,
-            "stable_delay_s": 0,
-        },
-    )
-    attempts = 0
-    handles: list[SandboxHandle] = []
-
-    async def fake_exec(
-        handle: SandboxHandle,
-        command: str,
-        *,
-        cwd: str | None = None,
-        env: dict[str, str] | None = None,
-        timeout_s: int | float | None = None,
-        user: str | int | None = None,
-    ) -> SandboxExecResult:
-        del command, cwd, env, timeout_s, user
-        nonlocal attempts
-        attempts += 1
-        handles.append(handle)
-        if attempts <= 2:
-            raise ConnectionError("direct execd endpoint is not accepting connections yet")
-        return SandboxExecResult(stdout="", stderr="", return_code=0)
-
-    monkeypatch.setattr(provider, "_exec", fake_exec)
-    handle = SandboxHandle(sandbox_id="sdk-sandbox-0", provider_name="opensandbox", raw=object())
-
-    await provider._verify_created_handle(handle)
-
-    assert attempts == 4
-    assert {seen_handle.sandbox_id for seen_handle in handles} == {"sdk-sandbox-0"}
-
-
-def test_opensandbox_create_probe_failures_are_retryable() -> None:
-    (
-        opensandbox_provider_module,
-        _OpenSandboxProvider,
-        OpenSandboxCreateVerificationError,
-        *_unused,
-    ) = _require_opensandbox_provider()
-
-    error = OpenSandboxCreateVerificationError("pod sdk-sandbox-0 failed create probe")
-
-    assert isinstance(error, SandboxCreateError)
-    assert opensandbox_provider_module._is_retryable_create_error(error) is True
-
-
-def test_opensandbox_starting_pod_endpoint_errors_are_retryable() -> None:
-    opensandbox_provider_module, *_unused = _require_opensandbox_provider()
-
-    error = RuntimeError(
-        "Get endpoint for sandbox sdk-sandbox-0 port 44772 failed: "
-        "Pod IP is not yet available. The Pod may still be starting."
-    )
-
-    assert opensandbox_provider_module._is_retryable_create_error(error) is True
-
-
-@requires_tenacity
-def test_opensandbox_exec_retries_retryable_sdk_failures(monkeypatch) -> None:
-    asyncio.run(_assert_opensandbox_exec_retries_retryable_sdk_failures(monkeypatch))
-
-
-async def _assert_opensandbox_exec_retries_retryable_sdk_failures(monkeypatch) -> None:
-    opensandbox_provider_module, OpenSandboxProvider, *_unused = _require_opensandbox_provider()
-
-    class FakeRunCommandOpts:
-        def __init__(self, **kwargs: Any) -> None:
-            self.kwargs = kwargs
-
-    class FakeLog:
-        def __init__(self, text: str) -> None:
-            self.text = text
-
-    class FakeLogs:
-        stdout = [FakeLog("ok")]
-        stderr: list[FakeLog] = []
-
-    class FakeExecution:
-        logs = FakeLogs()
-        error = None
-        exit_code = 0
-
-    class FakeCommands:
-        def __init__(self) -> None:
-            self.calls = 0
-
-        async def run(self, command: str, *, opts: FakeRunCommandOpts) -> FakeExecution:
-            del command, opts
-            self.calls += 1
-            if self.calls <= 2:
-                raise ConnectionError("transient connection failure")
-            return FakeExecution()
-
-    class FakeRaw:
-        def __init__(self) -> None:
-            self.commands = FakeCommands()
-
-    monkeypatch.setattr(
-        opensandbox_provider_module,
-        "_require_opensandbox_sdk",
-        lambda: (object, object, FakeRunCommandOpts, object, object),
-    )
-
-    provider = OpenSandboxProvider(
-        operations={
-            "retries": 2,
-            "retry_delay_s": 0,
-            "retry_max_delay_s": 0,
-            "command_retries": 2,
-        },
-        probe={"command": None},
-    )
-    raw = FakeRaw()
-    handle = SandboxHandle(sandbox_id="sdk-sandbox-1", provider_name="opensandbox", raw=raw)
-
-    result = await provider.exec(handle, "echo hello", timeout_s=30)
-
-    assert result.stdout == "ok"
-    assert result.return_code == 0
-    assert raw.commands.calls == 3
-
-
-@requires_tenacity
-def test_opensandbox_command_retries_default_to_disabled(monkeypatch) -> None:
-    asyncio.run(_assert_opensandbox_command_retries_default_to_disabled(monkeypatch))
-
-
-async def _assert_opensandbox_command_retries_default_to_disabled(monkeypatch) -> None:
-    opensandbox_provider_module, OpenSandboxProvider, *_unused = _require_opensandbox_provider()
-
-    class FakeRunCommandOpts:
-        def __init__(self, **kwargs: Any) -> None:
-            self.kwargs = kwargs
-
-    class FakeCommands:
-        def __init__(self) -> None:
-            self.calls = 0
-
-        async def run(self, command: str, *, opts: FakeRunCommandOpts) -> None:
-            del command, opts
-            self.calls += 1
-            raise ConnectionError("transient connection failure")
-
-    class FakeRaw:
-        def __init__(self) -> None:
-            self.commands = FakeCommands()
-
-    monkeypatch.setattr(
-        opensandbox_provider_module,
-        "_require_opensandbox_sdk",
-        lambda: (object, object, FakeRunCommandOpts, object, object),
-    )
-
-    provider = OpenSandboxProvider(
-        operations={
-            "retries": 2,
-            "retry_delay_s": 0,
-            "retry_max_delay_s": 0,
-        },
-        probe={"command": None},
-    )
-    raw = FakeRaw()
-    handle = SandboxHandle(sandbox_id="sdk-sandbox-1", provider_name="opensandbox", raw=raw)
-
-    try:
-        await provider.exec(handle, "echo hello", timeout_s=30)
-    except ConnectionError:
-        pass
-    else:
-        raise AssertionError("expected provider.exec to propagate the command failure")
-
-    assert raw.commands.calls == 1
-
-
-@requires_tenacity
-def test_opensandbox_close_timeout_does_not_fail_after_stop() -> None:
-    asyncio.run(_assert_opensandbox_close_timeout_does_not_fail_after_stop())
-
-
-async def _assert_opensandbox_close_timeout_does_not_fail_after_stop() -> None:
-    _opensandbox_provider_module, OpenSandboxProvider, *_unused = _require_opensandbox_provider()
-
-    class SlowCloseRaw:
-        def __init__(self) -> None:
-            self.killed = False
-
-        async def kill(self) -> None:
-            self.killed = True
-
-        async def close(self) -> None:
-            await asyncio.sleep(60)
-
-    raw = SlowCloseRaw()
-    provider = OpenSandboxProvider(
-        operations={"close_timeout_s": 0.01},
-        probe={"command": None},
-    )
-    handle = SandboxHandle(sandbox_id="sdk-sandbox-1", provider_name="opensandbox", raw=raw)
-
-    await provider.close(handle)
-
-    assert raw.killed is True
-
-
-@requires_tenacity
-def test_opensandbox_close_propagates_stop_failure() -> None:
-    asyncio.run(_assert_opensandbox_close_propagates_stop_failure())
-
-
-async def _assert_opensandbox_close_propagates_stop_failure() -> None:
-    _opensandbox_provider_module, OpenSandboxProvider, *_unused = _require_opensandbox_provider()
-
-    class StopFailureRaw:
-        async def kill(self) -> None:
-            raise RuntimeError("stop failed")
-
-        async def close(self) -> None:
-            return None
-
-    provider = OpenSandboxProvider(
-        operations={"close_timeout_s": 0.01},
-        probe={"command": None},
-    )
-    handle = SandboxHandle(sandbox_id="sdk-sandbox-1", provider_name="opensandbox", raw=StopFailureRaw())
-
-    with pytest.raises(RuntimeError, match="stop failed"):
-        await provider.close(handle)
 
 
 def test_mini_swe_sandbox_environment_owns_conda_setup(monkeypatch) -> None:
